requirements.txt
beautifulsoup4==4.11.1
ipython==8.2.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
requests==2.27.1
streamlit==1.8.1
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
pillow==9.0.1 
